offscreen-canvas/the-offscreen-canvas/2d.getcontext.exists.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.getcontext.exists</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.getcontext.exists</h1>
<p class="desc">The 2D context is implemented</p>


<script>
var t = async_test("The 2D context is implemented");
t.step(function() {
  var osc2D = "OffscreenCanvasRenderingContext2D";
  assert_own_property(globalThis, osc2D, osc2D + " is defined");
t.done();
});
</script>
